+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## main
 
+* Add back `migrate`, `validate` and `format` to the exported methods of this package
 * Add terrain to diff method and improve type. This also removes the `operations` from the API [#460](https://github.com/maplibre/maplibre-style-spec/pull/460)
 
 ## 19.3.3